fix 2 Multithreading homework
